Refactor transcript and video processing functionality
- Enhanced transcript retrieval in `transcript.py` with retry logic for better error handling.
- Increased thread pool size in `toDatabase.py` to improve processing efficiency.
- Modularized database interactions by introducing `DatabaseManager` in `db_manager.py`.
- Created `YouTubeAPIClient` in `youtube_client.py` to encapsulate YouTube API interactions.
- Implemented `TranscriptService` in `transcript_service.py` for managing transcript-related operations.
- Updated `VideoProcessingService` in `video_service.py` to streamline video data processing and database saving.
- Added data models in `data_models.py` for structured data handling.
- Improved logging for better traceability of operations across services.
