+++ api/__init__.py
@@ -1,0 +1,3 @@
+from .youtube_client import YouTubeAPIClient
+
+__all__ = ['YouTubeAPIClient']
