+++ database/__init__.py
@@ -1,0 +1,3 @@
+from .db_manager import DatabaseManager
+
+__all__ = ['DatabaseManager']
